Test(driver): 手書き fakeQueryRecorder を生成モックへ置換しテストを白箱/黒箱で再編
PR #488 のレビュー指摘に対応し、手書きの fakeQueryRecorder を削除して
生成モック mock_driver.MockQueryRecorder に置換する。

QueryRecorder/QueryAttrs は循環回避のため driver パッケージ内に定義されており、
その生成モック(mock_driver)は driver を import する。そのためモックを使う
テストはインパッケージ(package driver)では import 循環となり、外部テスト
パッケージ(package driver_test)に置く必要がある。これに合わせ、テストを
意味のある単位で再編する。

- 白箱(未公開ロジック)テストは package driver に維持:
  - query_metric_test.go: WithQueryName / classifyOperation / classifyErrorClass /
    buildQueryAttrs（fakeQueryRecorder の定義を削除）
  - query_tracer_test.go: NewQueryTracer の配線・TraceQueryEnd のログ分岐・
    endFields のマスク（recorder=nil 経路を検証）
- 黒箱(振る舞い)テストは package driver_test へ集約し生成モックを使用:
  - query_metric_external_test.go: 実 DB 経由で MockQueryRecorder により
    Observe 呼び出しと QueryAttrs(成功/失敗の status・error_class、ラベルへの
    SQL 本文非混入)を検証
  - transaction_retry_test.go: mock_driver を使う TestTxManager_Do_Retry を
    transaction_test.go から移設（未公開 defaultTxMaxAttempts は同値の
    dbCfg.TxMaxRetries() に置換）。これにより transaction_test.go(package driver)
    が mock_driver を import しなくなり、循環を解消する。
